Move updating of cb out of KNNSettings
Creates a new class called KNNCircuitBreakUtil in order to perform
updating of circuit break actions. Unable to move directly into
KNNCircuitBreaker due to race conditions that would be produced with the
NativeMemoryCacheManager. In the future, CB logic will be refactored as
well.
